% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/add.R
\docType{methods}
\name{add}
\alias{add}
\alias{add,GeneSetCollection,character,character-method}
\title{Add a relationship}
\usage{
add(object, gene, pathway)

\S4method{add}{GeneSetCollection,character,character}(object, gene = NULL,
  pathway = NULL)
}
\arguments{
\item{object}{A GeneSetCollection}

\item{gene}{A vector of character with genes.}

\item{pathway}{A vector of character with pathways.}
}
\value{
A GeneSetCollection with those relationships added.
}
\description{
Given a GeneSetCollection adds a relationship between a gene and a
pathway.
}
\details{
It accepts one gene and one pathway, several genes and one pathway and
several pathways and one gene, but not several pathways and several genes.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,gene = character,pathway = character}: Add relationships between genes and gene sets
}}

\seealso{
\code{\link{drop}} To just remove a gene or a pathways.
\code{\link{modify}} To remove and add relationship at once.
}
