% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{adjacency}
\alias{adjacency}
\title{Adjacency of the genes in GeneSetCollections}
\usage{
adjacency(object)
}
\arguments{
\item{object}{A \code{GeneSetCollection}}
}
\value{
A matrix
}
\description{
An adjacency matrix summarizes the connection between  genes.
The diagonal is set to 1.
}
\examples{
adjacency(Info)
}
\seealso{
\link[GSEABase]{incidence}
}
\author{
Lluís Revilla Sancho
}
