% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_GeneSetCollection.R
\docType{methods}
\name{as.GeneSetCollection}
\alias{as.GeneSetCollection}
\alias{as.GeneSetCollection,list,logical-method}
\alias{as.GeneSetCollection,list,missing-method}
\alias{as.GeneSetCollection,AnnDbBimap,ANY-method}
\title{Convert a list or a Bimap interface into GeneSetCollections}
\usage{
as.GeneSetCollection(object, filter, ...)

\S4method{as.GeneSetCollection}{list,logical}(object, filter, ...)

\S4method{as.GeneSetCollection}{list,missing}(object, filter, ...)

\S4method{as.GeneSetCollection}{AnnDbBimap,ANY}(object, filter, ...)
}
\arguments{
\item{object}{A list of genes and their pathways or an AnnDbBimap.}

\item{filter}{A logical}

\item{...}{Other unused parameters passed down.}
}
\value{
A GeneSetCollection
}
\description{
Transform the list or Bimap structure into a GeneSetCollection with unique
gene identifiers an pathnames.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = list,filter = logical}: Convert list to GeneSetCollections

\item \code{object = list,filter = missing}: Convert list to GeneSetCollections

\item \code{object = AnnDbBimap,filter = ANY}: Convert AnnDbBimap to GeneSetCollections
}}

\seealso{
\code{\link[GSEABase]{GeneSetCollection}}
}
\author{
Lluís Revilla
}
