% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-nPathways.R
\name{calc.max.nGenes}
\alias{calc.max.nGenes}
\title{Estimate the max number of genes}
\usage{
calc.max.nGenes(nPathways)
}
\arguments{
\item{nPathways}{Numeric value of number of pathways}
}
\value{
Numeric value of number of possible genes
}
\description{
Given the number of pathways calculates the max number of possible genes
}
\examples{
calc.max.nGenes(6)
}
