% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-nPathways.R
\name{calc.min.nGenes}
\alias{calc.min.nGenes}
\title{Estimate the min number of genes}
\usage{
calc.min.nGenes(nPathways)
}
\arguments{
\item{nPathways}{Numeric value of number of pathways}
}
\value{
Numeric value of number of possible genes
}
\description{
Given the number of pathways calculates the min number of possible genes
}
\examples{
calc.min.nGenes(6)
}
