% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-nPathways.R
\name{calc.nGenes}
\alias{calc.nGenes}
\title{Estimate the number of genes}
\usage{
calc.nGenes(nPathways)
}
\arguments{
\item{nPathways}{Numeric value of number of pathways}
}
\value{
Numeric value of number of possible genes
}
\description{
Given the number of pathways calculates the minimum number of possible genes
}
\examples{
calc.nGenes(6)
}
