% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-nPathways.R
\name{calc.nPathways}
\alias{calc.nPathways}
\title{Estimate the number of pathways}
\usage{
calc.nPathways(nGenes)
}
\arguments{
\item{nGenes}{Numeric value of number of genes}
}
\value{
Numeric value of number of possible pathways
}
\description{
Given the number of genes calculates the number of possible pathways
}
\examples{
calc.nPathways(6)
}
