% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/check-GSC.R
\docType{methods}
\name{check}
\alias{check}
\alias{check,GeneSetCollection-method}
\alias{geneIdType,GeneSetCollection-method}
\alias{collectionType,GeneSetCollection-method}
\title{Checks a GeneSetCollection}
\usage{
check(object)

\S4method{check}{GeneSetCollection}(object)

\S4method{geneIdType}{GeneSetCollection}(object)

\S4method{collectionType}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection}
}
\value{
A geneSetCollection
}
\description{
Checks that all the collection types is the same. Issues a warning when a
GOCollection is detected. Checks tat all the geneIdTypes is the same for
all the GeneSets. Checks that a GeneSet is bigger or equal to two genes.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Applies the checks

\item \code{GeneSetCollection}: Returns the geneIdType present in the GeneSetCollection

\item \code{GeneSetCollection}: Returns the collectionType present in the GeneSetCollection
}}

\examples{
isTRUE(check(Info))
data(sample.ExpressionSet)
ai <- AnnotationIdentifier(annotation(sample.ExpressionSet))
geneIds <- featureNames(sample.ExpressionSet)[100:109]
gs3 <- GeneSet(geneIds=geneIds, type=ai,
               setName="sample1", setIdentifier="102")
uprotIds <- c("Q9Y6Q1", "A6NJZ7", "Q9BXI6", "Q15035", "A1X283",
              "P55957")
gs4 <- GeneSet(uprotIds, geneIdType=UniprotIdentifier())
gsc <- GeneSetCollection(list(gs3, gs4))
gsc
\donttest{check(gsc)}
geneIdType(Info)
collectionType(Info)
}
