% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/keep.R
\docType{methods}
\name{combnGPP}
\alias{combnGPP}
\alias{combnGPP,GeneSetCollection-method}
\title{The combinations of genes per pathways}
\usage{
combnGPP(object)

\S4method{combnGPP}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A \code{GeneSetCollection}}
}
\value{
A numeric value
}
\description{
The number of combinations that can lead to the current
distribution of genes per pathway
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Estimate the combinations of genes per pathways
}}

\examples{
combnGPP(Info)
}
\seealso{
\code{\link{combnPPG}}
}
