% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/keep.R
\docType{methods}
\name{combnPPG}
\alias{combnPPG}
\alias{combnPPG,GeneSetCollection-method}
\title{The combinations of pathways per gene}
\usage{
combnPPG(object)

\S4method{combnPPG}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A \code{GeneSetCollection}}
}
\value{
A numeric value
}
\description{
The number of combinations that can lead to the current
distribution of pathways per genes
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Estimate the combinations of pathways per gene
}}

\examples{
combnPPG(Info)
}
\seealso{
\code{\link{combnGPP}}
}
\author{
Lluís Revilla Sancho
}
