% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nested.R
\docType{methods}
\name{compare}
\alias{compare}
\alias{compare,GeneSetCollection,GeneSetCollection-method}
\title{Pathways equal between two GeneSetCollections}
\usage{
compare(object1, object2)

\S4method{compare}{GeneSetCollection,GeneSetCollection}(object1, object2)
}
\arguments{
\item{object1, object2}{A GeneSetCollection}
}
\value{
The names of the pathways that are equivalent.
}
\description{
Check how many pathways are inside another geneSetCollection
}
\section{Methods (by class)}{
\itemize{
\item \code{object1 = GeneSetCollection,object2 = GeneSetCollection}: Check which pathways are described in both
GeneSetCollections
}}

