% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completness.R
\name{completness}
\alias{completness}
\title{Number of possible combinations}
\usage{
completness(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A named numeric vector
}
\description{
Calculates the number of combinations of each size.
}
\examples{
genesPerPathway <- c(2, 4, 5, 8, 5, 8)
completness(genesPerPathway)
}
