% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/condGene.R, R/condPath.R
\docType{methods}
\name{condGene}
\alias{condGene}
\alias{condGene,GeneSetCollection-method}
\alias{condPath,GeneSetCollection-method}
\title{Conditional probability of pathway length given the number of pathways}
\usage{
condGene(object)

\S4method{condGene}{GeneSetCollection}(object)

\S4method{condPath}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
A matrix
}
\description{
Conditional probability of pathway length given the number of pathways
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Given the number of pathways calculates the probability
of having a size of gene set

\item \code{GeneSetCollection}: Given the size of gene set calculates the probability
of having a number of pathways
}}

\author{
Lluís Revilla
}
