% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{condPath}
\alias{condPath}
\title{Conditional probability of the number of pathways given a pathway length}
\usage{
condPath(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
A matrix
}
\description{
Conditional probability of the number of pathways given a pathway length
}
\author{
Lluís Revilla
}
