% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cond.R
\docType{methods}
\name{condPerGenes}
\alias{condPerGenes}
\alias{condPerGenes,GeneSetCollection,missing-method}
\alias{condPerGenes,GeneSetCollection,numeric-method}
\title{Conditional probability by genes per pathway}
\usage{
condPerGenes(object, genesPerPathway)

\S4method{condPerGenes}{GeneSetCollection,missing}(object)

\S4method{condPerGenes}{GeneSetCollection,numeric}(object, genesPerPathway)
}
\arguments{
\item{object}{A GeneSetCollection}

\item{genesPerPathway}{A numeric vector of the number of genes per pathways.}
}
\value{
A matrix with the probabilities
}
\description{
Calculates the probability of having a number of pathways per gene given the
number of genes in a pathway
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,genesPerPathway = missing}: Check the conditional probability given the
number of genes

\item \code{object = GeneSetCollection,genesPerPathway = numeric}: Check the conditional probability given the
number of genes
}}

\examples{
condPerGenes(Info)
condPerGenes(Info, 2)
}
\seealso{
\code{\link{condPerPathways}}
}
\author{
Lluís Revilla Sancho
}
