% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cond.R
\docType{methods}
\name{condPerPathways}
\alias{condPerPathways}
\alias{condPerPathways,GeneSetCollection,missing-method}
\alias{condPerPathways,GeneSetCollection,numeric-method}
\title{Conditional probability by genes per pathway}
\usage{
condPerPathways(object, pathwaysPerGene)

\S4method{condPerPathways}{GeneSetCollection,missing}(object)

\S4method{condPerPathways}{GeneSetCollection,numeric}(object, pathwaysPerGene)
}
\arguments{
\item{object}{A GeneSetCollection}

\item{pathwaysPerGene}{A numeric vector of the number of pathways per genes.}
}
\value{
A matrix with the probabilities
}
\description{
Calculates the probability of having a number of genes per pathway given the
number of pathways in a gene
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,pathwaysPerGene = missing}: Check the conditional probability given the
number of pathways

\item \code{object = GeneSetCollection,pathwaysPerGene = numeric}: Check the conditional probability given the
number of pathways
}}

\examples{
condPerPathways(Info)
condPerPathways(Info, 2)
}
\seealso{
\code{\link{condPerGenes}}
}
\author{
Lluís Revilla Sancho
}
