% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/distr.R
\docType{methods}
\name{distrGenes}
\alias{distrGenes}
\alias{distrGenes,GeneSetCollection-method}
\title{Distributions of pathways in genes}
\usage{
distrGenes(object)

\S4method{distrGenes}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A \code{GeneSetCollection} object.}
}
\value{
\code{distrGenes} returns a matrix with the number of
pathways shared with other genes
}
\description{
Looks for the distribution of the pathways in other genes
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Calculates pathways shared with other genes of the
GeneSetCollection.
}}

\examples{
distrGenes(Info)
}
\seealso{
\code{\link{distrPathways}}
}
