% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/distr.R
\docType{methods}
\name{distrPathways}
\alias{distrPathways}
\alias{distrPathways,GeneSetCollection-method}
\title{Distributions of genes in pathways}
\usage{
distrPathways(object)

\S4method{distrPathways}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A \code{GeneSetCollection} object.}
}
\value{
\code{distrPathways} returns a matrix with the number of
genes shared with other pathways.
}
\description{
Looks for the distribution of the genes in other pathways.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Calculates genes shared with other pathways of the
GeneSetCollection.
}}

\examples{
distrPathways(Info)
}
\seealso{
\code{\link{distrGenes}}
}
