% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{double.factorial}
\alias{double.factorial}
\title{Double factorial}
\usage{
double.factorial(x)
}
\arguments{
\item{x}{The number of which the double factorial is wanted}
}
\value{
The double factorial of the number
}
\description{
Calculates the double factorial of the numbers, ie the product of all the
numbers that appear in the pair position.
}
\examples{
double.factorial(4)
double.factorial(5)
}
