% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{drop}
\alias{drop}
\title{Remove from a GeneSetCollection}
\usage{
drop(obj, gene = NULL, pathway = NULL)
}
\arguments{
\item{obj}{A GeneSetCollection object.}

\item{gene}{The number or names of the genes to be removed.}

\item{pathway}{The number or names of the pathways to be removed.}
}
\value{
A GeneSetCollection object without those genes or pathways.
}
\description{
Removes either genes or pathways of a GeneSetCollection if present.
}
\note{
Remember that a GeneSet is defined as a group of genes (more than one),
and if one gene set has only one gene it is removed. Conversely, a gene is
removed if it doesn't belong to any GeneSet.
}
