% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{dropRel}
\alias{dropRel}
\title{Remove a relationship from a GeneSetCollection}
\usage{
dropRel(object, gene, pathway)
}
\arguments{
\item{object}{A GeneSetCollection object.}

\item{gene}{The name of the gene.}

\item{pathway}{The name of the pathway.}
}
\value{
A GeneSetCollection object without that relationship.
}
\description{
Removes a gene-pathway relationship of a GeneSetCollection if present.
}
\note{
Remember that a GeneSet is defined as a group of genes (more than one),
and if one gene set has only one gene it is removed. Conversely, a gene is
removed if it doesn't belong to any GeneSet.
}
