% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/duplicated.R
\docType{methods}
\name{duplicatedGenes}
\alias{duplicatedGenes}
\alias{duplicatedGenes,GeneSetCollection-method}
\title{Duplicated genes}
\usage{
duplicatedGenes(object)

\S4method{duplicatedGenes}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection}
}
\value{
If some genes are duplicated returns \code{TRUE} else \code{FALSE}
}
\description{
Check if there are some pathways that are only present in the same genes.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Checks if different genes are described as
the same combination of pathways
}}

\examples{
duplicatedGenes(Info)
}
\seealso{
\code{\link{duplicatedPathways}}
}
