% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/edges.R
\docType{methods}
\name{edges}
\alias{edges}
\alias{nEdges}
\alias{nEdges,GeneSetCollection-method}
\alias{edges,GeneSetCollection-method}
\title{Edges between genes in GeneSetCollections}
\usage{
edges(object)

nEdges(object)

\S4method{nEdges}{GeneSetCollection}(object)

\S4method{edges}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A \code{GeneSetCollection}}
}
\value{
\code{edges} returns a \code{data.frame} where each row is an edge.

\code{nEdges} returns the number of edges between genes in
GeneSetCollection.
}
\description{
Edges return a data.frame with the names of the genes being connected.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Number of edges

\item \code{GeneSetCollection}: Genes between which there is an edge
}}

\examples{
edges(Info)
nEdges(Info)
}
\seealso{
\link{adjacency}, \link[GSEABase]{incidence}
}
\author{
Lluís Revilla Sancho
}
