% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{equalize}
\alias{equalize}
\title{Convert to characters}
\usage{
equalize(x)
}
\arguments{
\item{x}{A list with keys and values}
}
\value{
A list with two elements all the keys and all the values associated
}
\description{
Convert to characters
}
\author{
Lluís Revilla
}
\keyword{internal}
