% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sizes.R
\name{estimate.nGenes}
\alias{estimate.nGenes}
\title{Estimate range of genes}
\usage{
estimate.nGenes(gpp)
}
\arguments{
\item{gpp}{Numeric vector of genes per pathways}
}
\value{
A vector of the upper and lower extreme values of number of genes
that can create such combination
}
\description{
For a given vector of genes per pathways it calculates how many genes are able
to create this combination of sizes.
}
\examples{
estimate.nGenes(c(4, 5, 5, 5, 5, 5))
}
