% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sizes.R
\name{estimate.nPathways}
\alias{estimate.nPathways}
\title{Estimate range of pathways}
\usage{
estimate.nPathways(ppg)
}
\arguments{
\item{ppg}{A numeric vector of pathways per genes}
}
\value{
The range of possible number of pathways.
}
\description{
For a given vector of pathays per gene it calculcates how many pathways are
possible. Note, the minimum is always 1
}
\note{
If the number of genes is greater than 1024 it will return Inf for that
}
\examples{
estimate.nPathways(c(4, 5, 5, 5, 5, 5))
}
