% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate a GeneSetCollection}
\usage{
estimate(nGenes = NULL, nPathways = NULL, genesPerPathway = NULL,
  pathwaysPerGene = NULL, sizeGenes = NULL, sizePathways = NULL,
  sizePerGenes = NULL, sizePerPathways = NULL)
}
\arguments{
\item{nGenes}{Numeric value with the number of genes}

\item{nPathways}{Numeric value with the number of pathways}

\item{genesPerPathway}{Numeric vector with the number of genes per pathway}

\item{pathwaysPerGene}{Numeric vector with the number of pathways per gene}

\item{sizeGenes}{Numeric matrix with the size of pathways in genes see \code{\link{sizeGenes}}}

\item{sizePathways}{Numeric matrix with the size of genes in pathways see \code{\link{sizePathways}}}

\item{sizePerGenes}{A numeric vector with the amount of pathways of different size per genes see \code{\link{sizesPerGene}}}

\item{sizePerPathways}{A numeric vector with the amount of genes of different size per pathways see \code{\link{sizesPerPathway}}}
}
\value{
A \code{data.frame} with the max and minimum value for each estimate.
\code{NA} is returned if it cannot be estimated.
}
\description{
Given information about a \code{GeneSetCollection} it estimates the range of
the different parameters.
If the parameter should be a matrix, it only outputs the maximum and the
lowest value it could reach.
}
\examples{
estimate(nGenes = 100)
estimate(nPathways = 100)
estimate(nGenes = 100, nPathways = 100)
}
