% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{estimateGPP}
\alias{estimateGPP}
\title{Estimate genes per pathway combinations}
\usage{
estimateGPP(gsc)
}
\value{
A numeric value
}
\description{
Estimate the number of combinations that can lead to the current
distribution of pathways per genes
}
