% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{estimatePPG}
\alias{estimatePPG}
\title{Estimate pathway per gene combinations}
\usage{
estimatePPG(gsc)
}
\value{
A numeric value
}
\description{
Estimate the number of combinations that can lead to the current
distribution of pathways per genes
}
