% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGSC.R
\name{fromGPP}
\alias{fromGPP}
\title{Simulate GeneSetCollection}
\usage{
fromGPP(gpp)
}
\arguments{
\item{gpp}{The number of pathways per genes of the desired GeneSetCollection.}
}
\value{
A GeneSetCollection
}
\description{
Simulates a GeneSetCollection following the number of pathways per genes
introduced.
}
\seealso{
\code{\link{fromPPG}}
}
