% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGSC.R
\name{fromGPP_nGenes}
\alias{fromGPP_nGenes}
\title{Simulate a GeneSetCollection}
\usage{
fromGPP_nGenes(gpp, nGenes)
}
\arguments{
\item{gpp}{Genes per pathway.}

\item{nGenes}{Number of genes.}
}
\value{
A GeneSetCollection with those genes distributed in those pathways.
}
\description{
Creates a GeneSetCollection with the desired number of genes per pathway
given the number of genes.
}
\seealso{
\code{\link{fromPPG_nPathways}}
}
