% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate2GSC.R
\name{fromPPG_GPP}
\alias{fromPPG_GPP}
\alias{fromGPP_PPG}
\title{Simulate GeneSetCollection}
\usage{
fromPPG_GPP(ppg, gpp)
}
\arguments{
\item{ppg}{The number of pathways per genes of the desired GeneSetCollection.}

\item{gpp}{The number of genes per pathway of the desired GeneSetCollection.}
}
\value{
A GeneSetCollection
}
\description{
Simulates a GeneSetCollection following distributions of genes per pathway
and of pathways per gene.
}
