% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGSC.R
\name{fromPPG_nPathways}
\alias{fromPPG_nPathways}
\title{Simulate a GeneSetCollection}
\usage{
fromPPG_nPathways(ppg, nPathways)
}
\arguments{
\item{ppg}{Pathways per gene.}

\item{nPathways}{Number of genes.}
}
\value{
A GeneSetCollection with those genes distributed in those pathways.
}
\description{
Creates a GeneSetCollection with the desired number of pathways per gene
given the number of pathways.
}
\seealso{
\code{\link{fromGPP_nGenes}}
}
