% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromSizeGenes.R
\name{fromSizeGenes}
\alias{fromSizeGenes}
\title{Simulate a GeneSetCollection}
\usage{
fromSizeGenes(sizeGenes)
}
\arguments{
\item{sizeGenes}{The distribution of genes by size of the pathways they are in}
}
\value{
A GeneSetCollection with this genes and relationships
}
\description{
Given a matrix of the size of the pathways where a gene is involved it simulates
a GeneSetCollection
}
