% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromSizeGenes_sizePathways.R
\name{fromSizeGenes_sizePathways}
\alias{fromSizeGenes_sizePathways}
\title{Create a GeneSetCollection}
\usage{
fromSizeGenes_sizePathways(sizeGenes, sizePathways)
}
\arguments{
\item{sizeGenes}{A matrix result of \code{sizeGenes}.}

\item{sizePathways}{A matrix result of \code{sizePathways}.}
}
\value{
A GeneSetCollection object
}
\description{
Using the genes and to how big pathways thye are associated together with
the pathways and the number of genes each gene in the pathways is associated
it creates a new pathway.
}
\seealso{
\code{\link{sizeGenes}}, \code{\link{sizePathways}}
}
