% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromSizePathways.R
\name{fromSizePathways}
\alias{fromSizePathways}
\title{Simulate a GeneSetCollection}
\usage{
fromSizePathways(sizePathways)
}
\arguments{
\item{sizePathways}{The distribution of pathways by number of the pathways
their genes are in}
}
\value{
A GeneSetCollection with this genes and relationships
}
\description{
Given a matrix of the number of the pathways where a gene is involved it
simulates a GeneSetCollection
}
