% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAdv-package.R
\docType{data}
\name{genesKegg}
\alias{genesKegg}
\title{A GeneSetCollection}
\format{A GeneSetCollection}
\source{
\code{\link{as.GeneSetCollection(org.Hs.egPATH)}}
}
\usage{
genesKegg
}
\description{
A version of pathways of the KEGG database from the org.Hs.eg.db package
}
\keyword{datasets}
