% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/genesPerGene.R
\docType{methods}
\name{genesPerGene}
\alias{genesPerGene}
\alias{genesPerGene,GeneSetCollection,missing-method}
\alias{genesPerGene,GeneSetCollection,character-method}
\title{Number genes related to a gene}
\usage{
genesPerGene(object, genes, unique = TRUE)

\S4method{genesPerGene}{GeneSetCollection,missing}(object, unique = TRUE)

\S4method{genesPerGene}{GeneSetCollection,character}(object, genes,
  unique = TRUE)
}
\arguments{
\item{object}{A GeneSetCollection}

\item{genes}{A character vector of the genes.}

\item{unique}{A logical, to choose between unique genes or not.}
}
\value{
A vector with the number of genes related to each gene.
}
\description{
How many genes of different length a pathway is involved in.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,genes = missing}: Number of genes for all genes

\item \code{object = GeneSetCollection,genes = character}: Number of genes per gene selected
}}

\examples{
genesPerGene(Info)
genesPerGene(Info, unique = FALSE)
 genesPerGene(Info, unique = )
genesPerGene(Info, "10")
genesPerGene(Info, "10", unique = FALSE)
}
\seealso{
\code{\link{pathwaysPerGene}}
}
\author{
Lluís Revilla Sancho
}
