% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/genesPerPathway.R
\docType{methods}
\name{genesPerPathway}
\alias{genesPerPathway}
\alias{genesPerPathway,GeneSetCollection-method}
\title{Genes per pathway}
\usage{
genesPerPathway(object)

\S4method{genesPerPathway}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
A list of genes per pathway.
}
\description{
Calculates the genes per pathway of a GeneSetCollection
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Calculates genes per pathway of the
GeneSetCollection
}}

\seealso{
\code{\link{pathwaysPerGene}}
}
\author{
Lluís Revilla
}
