% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/genesSizePathways.R
\docType{methods}
\name{genesSizePathways}
\alias{genesSizePathways}
\alias{genesSizePathways,GeneSetCollection-method}
\title{Gene set size - gene relationship}
\usage{
genesSizePathways(object)

\S4method{genesSizePathways}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
A matrix
}
\description{
Calculates the number of gene sets of the same size a gene has.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Calculates the number of pathways with the
same size for each gene
}}

\author{
Lluís Revilla
}
