% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAdv.R
\docType{package}
\name{GSEAdv}
\alias{GSEAdv}
\alias{GSEAdv-package}
\title{GSEAdv: A package for analyzing gene sets.}
\description{
The GSEAdv package provides methods enhancing the GSEABase package
}
\section{GSEAdv methods}{

\itemize{
\item genesPerPathway
\item pathwaysPerGene
}
}

