% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizes.R
\name{helper_sizes}
\alias{helper_sizes}
\title{Helper for sizes}
\usage{
helper_sizes(data1, data2)
}
\arguments{
\item{data1}{The list of genes to pathways or of pathways to genes}

\item{data2}{The inverse of the other}
}
\value{
A matrix with the sizes of the data grouped
}
\description{
Calculates the sizes of the input
}
\keyword{internal}
