% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{IC}
\alias{IC}
\title{Calculate the information content or entropy}
\usage{
IC(x)
}
\arguments{
\item{x}{The values to calculate}
}
\value{
The numeric value of entropy
}
\description{
Calculate the information content or entropy
}
\author{
Lluís Revilla
}
\keyword{internal}
