% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/check-GSC.R
\docType{methods}
\name{independence}
\alias{independence}
\alias{independence,GeneSetCollection-method}
\title{Checks independence of genes}
\usage{
independence(object)

\S4method{independence}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks if all the genes are only present in one gene set
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Checks if a all the genes are in a single pathway
}}

\examples{
fl <- system.file("extdata", "Broad.xml", package = "GSEABase")
gss <- getBroadSets(fl)
independence(gss)
}
\seealso{
\code{\link{isolation}}
}
