% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAdv-package.R
\docType{data}
\name{Info}
\alias{Info}
\title{A GeneSetCollection}
\format{A GeneSetCollection}
\source{
Derived from genesKegg
}
\usage{
Info
}
\description{
A small GeneSetCollection to test and do examples with
}
\keyword{datasets}
