% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/check-GSC.R
\docType{methods}
\name{isolation}
\alias{isolation}
\alias{isolation,GeneSetCollection-method}
\title{Checks isolated GeneSets}
\usage{
isolation(object)

\S4method{isolation}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection}
}
\value{
Invisible a logical value
}
\description{
Checks if any gene set has all the genes only present in that gene set.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Checks if a gene set is isolated.
}}

