% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{maxIC}
\alias{maxIC}
\title{Calculate the maximal information content or entropy}
\usage{
maxIC(x)
}
\arguments{
\item{x}{The values to calculate}
}
\value{
The numeric value of max entropy that can be achieved
}
\description{
Calculate the maximal information content or entropy
}
\author{
Lluís Revilla
}
\keyword{internal}
