% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{modify}
\alias{modify}
\title{Modify a relationship}
\usage{
modify(object, gene, pathway)
}
\arguments{
\item{object}{A GeneSetCollection}

\item{gene}{A vector of character with genes.}

\item{pathway}{A vector of character with pathways.}
}
\value{
A GeneSetCollection with those relationships flipped.
}
\description{
Given a GeneSetCollection adds or remove a relationship between a gene and a
pathway.
}
\details{
It accepts one gene and one pathway, several genes and one pathway and
several pathways and one gene, but not several pathways and several genes.
}
\seealso{
\code{\link{drop}} To just remove a gene or a pathways
}
