% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{names_vec}
\alias{names_vec}
\title{Give a name to a vector}
\usage{
names_vec(x, pre)
}
\arguments{
\item{x}{The numeric vector}

\item{pre}{The prefix for the names of the vector}
}
\value{
the same vector with names
names_vec(1:3, "GS_")
}
\description{
Given a numeric vector sets names
}
\keyword{internal}
