% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nested.R
\docType{methods}
\name{nested}
\alias{nested}
\alias{nested,GeneSetCollection-method}
\title{Pathways included in other pathways}
\usage{
nested(object)

\S4method{nested}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection}
}
\value{
A matrix with 1 if they are included and 0 if not. 1
means that the pathways on the rows are in the ones of the columns
}
\description{
Check if all the genes of a pathway are inside another pathway.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Check if pathways are included in another pathway.
}}

\examples{
nested(Info)
}
\seealso{
\code{\link{compare}}
}
