% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/n_functions.R
\docType{methods}
\name{nGenes}
\alias{nGenes}
\alias{nGenes,GeneSetCollection-method}
\title{Number of genes}
\usage{
nGenes(object)

\S4method{nGenes}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
A number
}
\description{
Calculates the number of genes.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Number of different genes
}}

\author{
Lluís Revilla
}
