% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/n_functions.R
\docType{methods}
\name{nPathways}
\alias{nPathways}
\alias{nPaths}
\alias{nPathways,GeneSetCollection-method}
\title{Number of GeneSets}
\usage{
nPathways(object)

\S4method{nPathways}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
A number
}
\description{
Calculates the number of GeneSets
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Number of GeneSets
}}

\author{
Lluís Revilla
}
