% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/pathwaysPerGene.R
\docType{methods}
\name{pathwaysPerGene}
\alias{pathwaysPerGene}
\alias{pathwaysPerGene,GeneSetCollection-method}
\title{Pathways per gene}
\usage{
pathwaysPerGene(object)

\S4method{pathwaysPerGene}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
a list of pathways per genes
}
\description{
Calculates the pathways per gene of a GeneSetCollection. In set theory it is
called cardinality.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Pathways per gene in the GeneSetCollection
}}

\seealso{
\code{\link{genesPerPathway}}
}
\author{
Lluís Revilla
}
