% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sizes.R
\docType{methods}
\name{sizePathways}
\alias{sizePathways}
\alias{sizePathways,GeneSetCollection,missing-method}
\alias{sizePathways,GeneSetCollection,character-method}
\title{Number of pathways in genes of pathways}
\usage{
sizePathways(object, pathway)

\S4method{sizePathways}{GeneSetCollection,missing}(object)

\S4method{sizePathways}{GeneSetCollection,character}(object, pathway)
}
\arguments{
\item{object}{A GeneSetCollection}

\item{pathway}{A vector of character with pathways.}
}
\value{
A matrix or a list of vectors if only some pathways are requested
}
\description{
Calculates the number of pathways per gene in pathways.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,pathway = missing}: Number of pathways per gene in all pathways

\item \code{object = GeneSetCollection,pathway = character}: Number of pathways per gene in all pathways
}}

\examples{
sizePathways(Info)
sizePathways(Info, "1430728")
sizePathways(Info, c("1430728", "156580"))
}
\seealso{
\code{\link{sizeGenes}} The equivalent for genes
}
