% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sizesPer.R
\docType{methods}
\name{sizesPerGene}
\alias{sizesPerGene}
\alias{sizesPerGene,GeneSetCollection,missing-method}
\alias{sizesPerGene,GeneSetCollection,character-method}
\title{Number of different size of pathways per gene}
\usage{
sizesPerGene(object, genes)

\S4method{sizesPerGene}{GeneSetCollection,missing}(object)

\S4method{sizesPerGene}{GeneSetCollection,character}(object, genes)
}
\arguments{
\item{object}{A GeneSetCollection}

\item{genes}{A character gene.}
}
\value{
A vector with the number of different pathways sizes of the genes.
}
\description{
How many pathways of different length a gene is involved in.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,genes = missing}: Number of sizes of pathways per all the genes

\item \code{object = GeneSetCollection,genes = character}: Number of sizes of pathways per the provided genes
}}

\seealso{
\code{\link{sizesPerPathway}}
}
\author{
Lluís Revilla Sancho
}
