% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sizesPer.R
\docType{methods}
\name{sizesPerPathway}
\alias{sizesPerPathway}
\alias{sizesPerPathway,GeneSetCollection,missing-method}
\alias{sizesPerPathway,GeneSetCollection,character-method}
\title{Number of different sizes of gene per pathway}
\usage{
sizesPerPathway(object, pathways)

\S4method{sizesPerPathway}{GeneSetCollection,missing}(object)

\S4method{sizesPerPathway}{GeneSetCollection,character}(object, pathways)
}
\arguments{
\item{object}{A GeneSetCollection}

\item{pathways}{A character vector of the pathways-}
}
\value{
A vector with the number of different gene sizes of the pathays
}
\description{
How many genes of different length a pathway is involved in.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,pathways = missing}: Number of sizes of genes per all pathways.

\item \code{object = GeneSetCollection,pathways = character}: Number of sizes of genes per pathways provided
}}

\seealso{
\code{\link{sizesPerGene}}
}
\author{
Lluís Revilla Sancho
}
