% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGSC.R
\name{w}
\alias{w}
\title{Simulate a GeneSetCollection}
\usage{
w(ppg, nPathways)
}
\arguments{
\item{ppg}{Pathways per gene.}

\item{nPathways}{Number of genes.}
}
\value{
A GeneSetCollection with those genes distributed in those pathways.
}
\description{
Creates a GeneSetCollection with the desired number of pathways per gene
given the number of pathways.
}
\seealso{
\code{\link{w}}
}
