% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGSC.R
\name{z}
\alias{z}
\title{Simulate GeneSetCollection}
\usage{
z(gpp)
}
\arguments{
\item{gpp}{The number of pathways per genes of the desired GeneSetCollection}
}
\value{
A GeneSetCollection
}
\description{
Simulates a GeneSetCollection following a distribution
}
