% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{compare}
\alias{compare}
\title{Pathways equal between two GeneSetCollections}
\usage{
compare(object1, object2)
}
\arguments{
\item{object1, object2}{A GeneSetCollection}
}
\value{
The names of the pathways that are equivalent.
}
\description{
Check how many pathways are inside another geneSetCollection
}
\examples{
compare(Info, genesKegg)
}
