% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{condPath}
\alias{condPath}
\title{Conditional probability of the number of pathways given pathways length}
\usage{
condPath(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
A matrix
}
\description{
Calculates the probability of a gene to have a number of pathways given the
size of pathways.
}
\details{
Uses \code{\link{check}} to validate the output.
}
\seealso{
\code{\link{condGene}}
}
\author{
Lluís Revilla
}
