% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/remove.R
\docType{methods}
\name{drop}
\alias{drop}
\alias{drop,GeneSetCollection-method}
\alias{dropRel,GeneSetCollection,character,character-method}
\title{Remove from a GeneSetCollection}
\usage{
drop(object, gene, pathway)

\S4method{drop}{GeneSetCollection}(object, gene = NULL, pathway = NULL)

\S4method{dropRel}{GeneSetCollection,character,character}(object, gene = NULL,
  pathway = NULL)
}
\arguments{
\item{object}{A GeneSetCollection object.}

\item{gene}{The number or names of the genes to be removed.}

\item{pathway}{The number or names of the pathways to be removed.}
}
\value{
A GeneSetCollection object without those genes or pathways.
}
\description{
Removes either genes or pathways of a GeneSetCollection if present
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Drop a gene a pathway or both

\item \code{object = GeneSetCollection,gene = character,pathway = character}: Drop a gene a pathway or both
}}

\note{
Remember that a GeneSet is defined as a group of genes (more than one),
and if one gene set has only one gene it is removed. Conversely, a gene is
removed if it doesn't belong to any GeneSet.
}
