% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/remove.R
\docType{methods}
\name{drop}
\alias{drop}
\alias{drop,GeneSetCollection,character,character-method}
\alias{drop,GeneSetCollection,missing,character-method}
\alias{drop,GeneSetCollection,character,missing-method}
\alias{drop,GeneSetCollection,numeric,numeric-method}
\alias{drop,GeneSetCollection,missing,numeric-method}
\alias{drop,GeneSetCollection,numeric,missing-method}
\title{Remove from a GeneSetCollection}
\usage{
drop(object, gene, pathway)

\S4method{drop}{GeneSetCollection,character,character}(object, gene, pathway)

\S4method{drop}{GeneSetCollection,missing,character}(object, pathway)

\S4method{drop}{GeneSetCollection,character,missing}(object, gene)

\S4method{drop}{GeneSetCollection,numeric,numeric}(object, gene, pathway)

\S4method{drop}{GeneSetCollection,missing,numeric}(object, pathway)

\S4method{drop}{GeneSetCollection,numeric,missing}(object, gene, pathway)
}
\arguments{
\item{object}{A GeneSetCollection object.}

\item{gene}{The number or names of the genes to be removed.}

\item{pathway}{The number or names of the pathways to be removed.}
}
\value{
A GeneSetCollection object without those genes or pathways.
}
\description{
Removes either genes or pathways of a GeneSetCollection if present.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,gene = character,pathway = character}: Drop genes and pathways

\item \code{object = GeneSetCollection,gene = missing,pathway = character}: Drop pathways

\item \code{object = GeneSetCollection,gene = character,pathway = missing}: Drop genes

\item \code{object = GeneSetCollection,gene = numeric,pathway = numeric}: Drop the specified number of genes and pathways

\item \code{object = GeneSetCollection,gene = missing,pathway = numeric}: Drop the specified number of pathway

\item \code{object = GeneSetCollection,gene = numeric,pathway = missing}: Drop the specified number of genes.
}}

\note{
Remember that a GeneSet is defined as a group of genes (more than one),
and if one gene set has only one gene it is removed. Conversely, a gene is
removed if it doesn't belong to any GeneSet.
}
\examples{
drop(Info, gene = "3", pathway = "156580")
drop(Info, pathway = "156580")
}
\seealso{
\code{\link{add}} To just add a relationship.
\code{\link{modify}} To remove and add relationship at once.
}
