% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/remove.R
\docType{methods}
\name{dropRel}
\alias{dropRel}
\alias{dropRel,GeneSetCollection,character,character-method}
\title{Remove a relationship from a GeneSetCollection}
\usage{
dropRel(object, gene, pathway)

\S4method{dropRel}{GeneSetCollection,character,character}(object, gene, pathway)
}
\arguments{
\item{object}{A GeneSetCollection object.}

\item{gene}{The number or names of the genes to be removed.}

\item{pathway}{The number or names of the pathways to be removed.}
}
\description{
Removes a relationship between genes and a pathway of a GeneSetCollection
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,gene = character,pathway = character}: a relationship between a genes and a pathway
}}

\examples{
genesPerPathway(Info)
out <- dropRel(Info, "9", "156580")
genesPerPathway(out)
}
\seealso{
\code{\link{add}} To just add a relationship.
\code{\link{modify}} To remove and add relationship at once.
\code{\link{drop}} To remove genes and pathways.
}
