% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/duplicated.R
\docType{methods}
\name{duplicatedPathways}
\alias{duplicatedPathways}
\alias{duplicatedPathways,GeneSetCollection-method}
\title{Duplicated pathways}
\usage{
duplicatedPathways(object)

\S4method{duplicatedPathways}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection}
}
\value{
If some pathways are duplicated returns \code{TRUE} else \code{FALSE}
}
\description{
Check if there are some genes that are only present in the same pathways.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Checks if a different pathways are described
as the same combination of genes.
}}

\examples{
duplicatedPathways(Info)
}
\seealso{
\code{\link{duplicatedGenes}}
}
