% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGSC.R
\name{fromPPG}
\alias{fromPPG}
\title{Simulate GeneSetCollection}
\usage{
fromPPG(ppg)
}
\arguments{
\item{ppg}{The number of pathways per genes of the desired GeneSetCollection}
}
\value{
A GeneSetCollection
}
\description{
Simulates a GeneSetCollection following a distribution
}
\note{
It might take some iterations to reach the desired output
}
\examples{
fromPPG(c(2, 2, 3, 2, 2, 2, 2))
}
\seealso{
\code{\link{fromGPP}}
}
