% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/modify.R
\docType{methods}
\name{modify}
\alias{modify}
\alias{drop,GeneSetCollection-method}
\title{Modify a relationship}
\usage{
modify(object, gene, pathway)

\S4method{drop}{GeneSetCollection}(object, gene = NULL, pathway = NULL)
}
\arguments{
\item{object}{A GeneSetCollection}

\item{gene}{A vector of character with genes.}

\item{pathway}{A vector of character with pathways.}
}
\value{
A GeneSetCollection with those relationships flipped.
}
\description{
Given a GeneSetCollection adds or remove a relationship between a gene and a
pathway.
}
\details{
It accepts one gene and one pathway, several genes and one pathway and
several pathways and one gene, but not several pathways and several genes.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Drop a gene a pathway or both
}}

\seealso{
\code{\link{drop}} To just remove a gene or a pathways
}
