% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{modify}
\alias{modify}
\title{Modify a relationship}
\usage{
modify(obj, gene, pathway)
}
\arguments{
\item{gene}{A vector of character with genes.}

\item{pathway}{A vector of character with pathways.}
}
\value{
A GeneSetCollection with those relationships turned
}
\description{
Given a GeneSetCollection adds or remove a relationship between a gene and a
pathway.
}
