% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/modify.R
\docType{methods}
\name{modify}
\alias{modify}
\alias{modify,GeneSetCollection,character,character-method}
\title{Modify a relationship}
\usage{
modify(object, gene, pathway)

\S4method{modify}{GeneSetCollection,character,character}(object, gene = NULL,
  pathway = NULL)
}
\arguments{
\item{object}{A GeneSetCollection}

\item{gene}{A vector of character with genes.}

\item{pathway}{A vector of character with pathways.}
}
\value{
A GeneSetCollection with those relationships flipped.
}
\description{
Given a GeneSetCollection adds or remove a relationship between a gene and a
pathway.
}
\details{
It accepts one gene and one pathway, several genes and one pathway and
several pathways and one gene, but not several pathways and several genes.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,gene = character,pathway = character}: Add or remove relationships between genes and gene sets
}}

\examples{
modify(Info, "Gene2", "156580")
modify(Info, c("Gene2", "Gene3"), "156580")
modify(Info, "Gene2", c("156580", "211859"))
}
\seealso{
\code{\link{drop}} To just remove a gene or a pathways
\code{\link{add}} To just add a relationship.
}
