% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/pathwaysPerGene.R
\docType{methods}
\name{pathwaysPerGene}
\alias{pathwaysPerGene}
\alias{pathwaysPerGene,GeneSetCollection,missing-method}
\alias{pathwaysPerGene,GeneSetCollection,character-method}
\title{Pathways per gene}
\usage{
pathwaysPerGene(object, genes)

\S4method{pathwaysPerGene}{GeneSetCollection,missing}(object)

\S4method{pathwaysPerGene}{GeneSetCollection,character}(object, genes)
}
\arguments{
\item{object}{A GeneSetCollection object.}

\item{genes}{A character vector of genes.}
}
\value{
A numeric vector of pathways per genes.
}
\description{
Calculates the pathways per gene of a GeneSetCollection. In set theory it is
called cardinality.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,genes = missing}: Pathways per gene in all the GeneSetCollection

\item \code{object = GeneSetCollection,genes = character}: Pathways per gene in the GeneSetCollection for
those genes.
}}

\examples{
pathwaysPerGene(Info)
pathwaysPerGene(Info, "10")
}
\seealso{
\code{\link{genesPerPathway}}
}
\author{
Lluís Revilla
}
