% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sizes.R
\docType{methods}
\name{sizeGenes}
\alias{sizeGenes}
\alias{sizeGenes,GeneSetCollection,missing-method}
\alias{sizeGenes,GeneSetCollection,character-method}
\title{Number of genes in pathways of genes}
\usage{
sizeGenes(object, gene)

\S4method{sizeGenes}{GeneSetCollection,missing}(object)

\S4method{sizeGenes}{GeneSetCollection,character}(object, gene)
}
\arguments{
\item{object}{A GeneSetCollection}

\item{gene}{A vector of character with genes.}
}
\value{
A matrix or a list of vectors if only some genes are requested
}
\description{
Calculates the number of pathways per gene in pathways.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,gene = missing}: Number of genes per pathway in all genes

\item \code{object = GeneSetCollection,gene = character}: Number of genes per pathway in all genes
}}

\examples{
sizeGenes(Info)
sizeGenes(Info, "2")
sizeGenes(Info, c("2", "3"))
}
\seealso{
\code{\link{sizePathways}} To equivalent for pathways
}
